+++ src/imio/helpers/patches.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-
+from imio.helpers.cache import invalidate_cachekey_volatile_for
+from Products.PlonePAS.plugins.role import GroupAwareRoleManager
 from smtplib import SMTP_SSL
 from zope.sendmail.mailer import SMTPMailer
 
@@ -29,3 +30,15 @@ def ssl_makeMailer(self):
     if self.smtp_port in (465, '465'):
         mailer.smtp = SMTP_SSL
     return mailer
+
+
+def assignRolesToPrincipal(self, roles, principal_id, REQUEST=None):  # noqa
+    GroupAwareRoleManager._old_assignRolesToPrincipal(self, roles, principal_id, REQUEST)
+    # we need to invalidate cachekey
+    invalidate_cachekey_volatile_for('_users_groups_value')
+
+
+def assignRoleToPrincipal(self, role_id, principal_id, REQUEST=None):
+    GroupAwareRoleManager._old_assignRoleToPrincipal(self, role_id, principal_id, REQUEST)
+    # we need to invalidate cachekey
+    invalidate_cachekey_volatile_for('_users_groups_value')